Added validation functionality to reflect what we have implemented  in LoginForm.
LinearProgress module has been imported to enchance and notify the user
 while they wait for the server to respond.

The typescript interface RegisterFormState was added to help type check
more state that was implemented for validation purposes.

useEffect life cycle method was implemented to focus on the first input field
when the component is mounted.

useRefs were added so that we can focus on them when a input field returns
an error.

State was refactored into one object and is now accessable via values. Set state
 is also available via setValues.

The _registerBtn method was enhanced with muitiple checks for empty strings on
mulitple fields. Runs the validateRegisteredUser method. Finally resets the
state if no errors are recieved.