Modifed axios to use params object and added return data.
The axios get call to the movie api was hard coding the api key and query straight into the url.

Although techinically this is the desired effect we are now passing it through the built in param object inside the second arugment of axios.

We are also no returning the results from teh data returned data object directly.